Fix failing tests + minor cleanup, use only one prop to decide whether to register mbeans or not Fixing tests: 	- enable statistics by default 	- add no-op statistics impl 	- account for sampled threads 	- account for default event listeners attached for put/update/remove/expire stats
git-svn-id: http://svn.terracotta.org/svn/ehcache@1095 b9324663-ca0f-0410-8574-be9b3887307d

+++ branches/ehcache-1.7-tc/core/src/main/java/net/sf/ehcache/statistics/CacheUsageListener.java
@@ -78,7 +78,7 @@ public interface CacheUsageListener {
      * 
      * @param millis
      */
-    public void notifyTimeTakenForGet(final int millis);
+    public void notifyTimeTakenForGet(final long millis);
 
     /**
      * Called when an element is expired in the cache
