+++ branches/ehcache-1.7-tc/core/src/main/java/net/sf/ehcache/statistics/CacheUsageStatisticsData.java
@@ -62,7 +62,7 @@ public interface CacheUsageStatisticsData extends CacheEventListener {
      * 
      * @param millis
      */
-    public void addGetTimeMillis(final int millis);
+    public void addGetTimeMillis(final long millis);
 
     /**
      * Called when an element is expired in the cache
